refactor: move exception handling logic to @ControllerAdvice for better separation
