+++ 04-spring-boot-rest-crud/01-spring-boot-rest-crud/src/main/java/np/com/krishnabk/demo/rest/StudentRestController.java
@@ -50,35 +50,4 @@ public class StudentRestController {
 
         return theStudents.get(studentId);
     }
-
-    // Add an exception handler using @ExceptionHandler
-    @ExceptionHandler
-    public ResponseEntity<StudentErrorResponse> handleException(StudentNotFoundException exception){
-
-        // create a StudentErrorResponse
-
-        StudentErrorResponse errorResponse = new StudentErrorResponse();
-        errorResponse.setStatus(HttpStatus.NOT_FOUND.value());
-        errorResponse.setMessage(exception.getMessage());
-        errorResponse.setTimeStamp(System.currentTimeMillis());
-
-        // return ResponseEntity
-        return new ResponseEntity<>(errorResponse, HttpStatus.NOT_FOUND);
-    }
-
-    // Add another exception handler ... to catch any exception (catch all)
-
-    @ExceptionHandler
-    public ResponseEntity<StudentErrorResponse> handleException(Exception e){
-
-        // create a StudentErrorResponse
-
-        StudentErrorResponse errorResponse = new StudentErrorResponse();
-        errorResponse.setStatus(HttpStatus.BAD_REQUEST.value());
-        errorResponse.setMessage(e.getMessage());
-        errorResponse.setTimeStamp(System.currentTimeMillis());
-
-        // return ResponseEntity
-        return new ResponseEntity<>(errorResponse, HttpStatus.BAD_REQUEST);
-    }
 }
